WIP: scc creates dfxp with <span> tags that don't close well.
Also should redo the InterpretableNodeCreator.add_chars method. It's
quite buggy now because it adds chars to the last text node... even if a
line break was specified after this text node. Should instead only add
chars to the last created node if it's a text node, or add chars to a
newly created text node.